[Refactor] [FE-10] 검색모달 리팩토링
